lights: Adjust RGB values when blinking (binary RGB mode)
When blinking, R, G and B parts can be either on or off. The parts
can't have variable brightness any more (unlike in the 'always on'
mode) as pwm is used for the blinking and is no longer available
for achieving different brightness of different RGB components to
produce something close to 2^24 different colors.

Reflect this limitation by adjusting the RGB values when blinking
to provide color results as close as possible to what user has
selected via the color picker, while staying limited to only
8 possible colors (2^3) allowed by the 3-bit binary RGB coding.

Change-Id: I9737d835ce5be148d7be68331b5c3a9f5d4b3f17